Refactor navigation and improve map functionality
- Move Map screen to secondary stack, remove from bottom tabs
- Fix GeoJSON layer not updating on route switch (Web)
- Fix user location marker visibility and updates (Web)
- Refactor tab screens to `app/(tabs)/`
