+++ app/(tabs)/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import HomeScreen from '../components/screens/HomeScreen';
+import HomeScreen from '../../components/screens/HomeScreen';
 
 export default function Layout(): React.ReactElement {
   return <HomeScreen />;
